+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Avoid calling `login` for host
+  [cyberark/conjur-api-java#117](https://github.com/cyberark/conjur-api-java/pull/117)
 
 ## [3.0.4] - 2023-02-27